THis code has added the ability to write the receipts
This allows us to have an archive of all the receipts once they have been finalized. The program then has a list of all the file names for future use.

This code is only able to create a new record of an order for a table every minute. This is a reasonable real world limitation, as in the real world it would be difficult for a table to serve two customers in less than a minute.